# swagger/api.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: ft_transcendence API
  description: API Document for ft_transcendence
servers:
  - url: "http://localhost:3001"

tags:
  - name: health
    description: ヘルスチェック
  - name: user
    description: ユーザー
  - name: chatRoom
    description: チャットルーム
  - name:  chat-block-user
    description: チャット機能でのミュートユーザ情報
  - name: chatRoomMembers
    description: チャットルームの参加者情報
  - name: match
    description: 進行中のマッチリスト

paths:
  /health:
    get:
      tags: 
        - health
      summary: Health check
      description: ヘルスチェック
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/health'

  /user:
    post:
      tags:
        - user
      summary: user
      description: ユーザーの作成
      requestBody:
        description: create a new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_create_req'
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user_create_res'

  /user/users:
    get:
      tags:
        - user
      summary: 全ユーザーの取得
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users_get_res'


  /user/{userID}:
    get:
      tags:
        - user
      summary: get user by ID
      description: IDを指定してuser情報を取得
      parameters:
        - name: userID
          required: true
          description: id
          in: path
          schema:
            type: integer
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user_get_res'
    delete:
      tags:
        - user
      summary: delete user
      parameters:
        - name: userID
          required: true
          description: id
          in: path
          schema:
            type: integer
      responses:
        200:
          content:
            text/plain:
              schema:
                type: string
                example: OK
    post:
      tags:
        - user
      summary: update user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#components/schemas/user_update_req'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user_get_res'

  /user/sign_in:
    post:
      tags:
        - user
      summary: Sign In
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#components/schemas/user_signin_req'
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user_get_res'
        401:
          description: 失敗 

  /user/friends/{userID}:
    get:
      tags:
        - user
      summary: get friends
      description: パラメータ(userID)のフレンドを取得
      parameters:
        - name: userID
          required: true
          in: path
          schema:
            type: integer
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/friends_get_res'

  /user/friends:
    post:
      tags:
        - user
      summary: request or accept friendship
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/friends_post_req'
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/friends_post_req'

  /user/friends/pending/{userID}:
    get:
      tags:
        - user
      summary: get pending list
      description: パラメータ(userID)のフレンド待機リストを取得
      parameters:
        - name: userID
          required: true
          in: path
          schema:
            type: integer
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/friends_get_res'

  /user/user_avatar/{userID}:
    get:
      tags:
        - user
      summary: get user avatar
      parameters:
        - name: userID
          required: true
          in: path
          schema:
            type: integer
      responses:
        200:
          description: 成功
          content:
            text/plain:
              schema:
                type: string
                example: hogehogebase64

  /chatroom:
    get:
      tags:
        - chatRoom
      summary: get chatRoom List
      description: chatRoom情報の一覧を取得
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/chat_room_res'
    post:
      summary: get chatRoom
      tags:
        - chatRoom
      description: chatRoom情報を新規作成
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chat_room_req'
      responses:
        201:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chat_room_res'

  /chatroom/{roomId}:
    post:
      summary: update chatRoom
      tags:
        - chatRoom
      description: chatRoom情報を更新
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chat_room_req'
      responses:
        201:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chat_room_res'

    delete:
      tags:
        - chatRoom
      summary: delete chatRoom
      description: chatRoom情報を削除
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/chat_room_res'
  /chatroom/{roomId}/auth:
    post:
      tags:
        - chatRoom
      summary: auth protected chatRoom
      description: chatRoomへのパスワード認証
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chat_room_auth_req'
      responses:
        201:
          description: 成功
        401:
          description: 認証失敗
  /chat-block-user:
    get:
      tags:
        - chat-block-user
      summary: get chat-block-user
      description: チャット機能でのミュートユーザ情報一覧を取得
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/chat-block-user'
    post:
      summary: update block
      tags:
        - chat-block-user
      description: チャット機能でのミュートユーザ情報を更新
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chat-block-user'
      responses:
        201:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chat-block-user'
    delete:
      summary: update block
      tags:
        - chat-block-user
      description: チャット機能でのミュートユーザ情報を削除
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chat-block-user-pk'
      responses:
        200:
          description: 成功
  /chat-block-user/{blockUserId}:
    get:
      tags:
        - chat-block-user
      summary: get block for the specified user
      description: チャット機能でのミュートユーザ情報の取得
      parameters:
        - in: path
          name: blockUserId
          required: true
          schema:
            type: number
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/chat-block-user'
  /chatRoomMembers:
    get:
      tags:
        - chatRoomMembers
      summary: get chatRoomMembers
      description: チャット機能でのルームの参加者情報を取得
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/chatRoomMember'
    post:
      summary: update chatRoomMembers
      tags:
        - chatRoomMembers
      description: チャット機能でのルームの参加者情報を更新
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chatRoomMember'
      responses:
        201:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chatRoomMember'
    delete:
      summary: delete chatRoomMembers
      tags:
        - chatRoomMembers
      description: チャット機能でのルームの参加者情報を削除
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chatRoomMemberPK'
      responses:
        200:
          description: 成功

  /match/matches:
    get:
      tags:
        - match
      summary: 全マッチの取得
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/matches_get'

  /match:
    post:
      tags:
        - match
      summary: match
      description: マッチの作成
      requestBody:
        description: create a new match
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/match_get'
      responses:
        200:
          description: 成功

  /match/{matchID}:
    get:
      tags:
        - match
      summary: get match by ID
      description: IDを指定してmatch情報を取得
      parameters:
        - name: matchID
          required: true
          description: id
          in: path
          schema:
            type: integer
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/match_get'

  /match/result/{matchID}:
    post:
      tags:
        - match
      summary: match result
      description: マッチの結果
      requestBody:
        description: Post match results
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/match_result_post'
      responses:
        200:
          description: 成功

components:
  schemas:
    health:
      type: object
      properties:
        health:
          type: string
          example: OK

    user_create_req:
      type: object
      properties:
        name:
          type: string
          example: Tom Brady
        password:
          type: string
          example: password
        avatar:
          type: string
          example: iVBORw0KGgBase64...

    user_create_res:
      type: object
      properties:
        id:
          type: integer
          example: 1

    user_update_req:
      type: object
      properties:
        name:
          type: string
          example: Updated!!!
        password:
          type: string
          example: update

    users_get_res:
      type: array
      items:
        $ref: '#/components/schemas/user_get_res'

    user_get_res:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Tom Brady

    user_signin_req:
      type: object
      properties:
        id:
          type: integer
          example: 1
        password:
          type: string
          example: password

    friends_get_res:
      type: array
      items:
        type: object
        properties:
          id:
            type: integer
            example: 1
          name:
            type: string
            example: Nick Foles
    
    friends_post_req:
      type: object
      properties:
        from:
          type: integer
          example: 1
        to:
          type: integer
          example: 2

    chat_room_res:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Tom Brady
        created_by_user_id:
          type: number
          example: 1
        public_id:
          type: string
          example: 'public'
    chat_room_req:
      type: object
      properties:
        name:
          type: string
          example: Tom Brady
        created_by_user_id:
          type: number
          example: 1
        public_id:
          type: string
          example: 'public'
    chat_room_auth_req:
      type: object
      properties:
        password:
          type: string
          example: "password"
    chat-block-user:
      type: object
      properties:
        blockUserId:
          type: number
          example: 1
        blockedUserId:
          type: number
          example: 2
        block_until:
          type: string
          example: "2024-01-31T14:59:00.000Z"
    chat-block-user-pk:
      type: object
      properties:
        blockUserId:
          type: number
          example: 1
        blockedUserId:
          type: number
          example: 2
    chatRoomMember:
      type: object
      properties:
        chatRoomId:
          type: number
          example: 1
        userId:
          type: number
          example: 1
        isBanned:
          type: boolean
          example: false
        isAdministrator:
          type: boolean
          example: false
    chatRoomMemberPK:
      type: object
      properties:
        chatRoomId:
          type: number
          example: 1
        userId:
          type: number
          example: 1

    match_get:
      type: object
      properties:
        id:
          type: number
          example: 0
        p1:
          type: number
          example: 1
        p2:
          type: number
          example: 1
        winner:
          type: number
          example: 1

    matches_get:
      type: array
      items:
        $ref: '#/components/schemas/match_get'

    match_result_post:
      type: object
      properties:
        id:
          type: number
          example: 1
        winner:
          type: number
          example: 1